Mc2.CrusTest.Application/Cutomers/Handlers/GetCustomerByIdQueryHandler.cs
using Mc2.CrudTest.Domain.Entities;
using Mc2.CrudTest.Domain.Interfaces;
using Mc2.CrusTest.Application.Cutomers.Queries;
using MediatR;

namespace Mc2.CrusTest.Application.Customers.Handlers
{
    public class GetCustomerByIdQueryHandler : IRequestHandler<GetCustomerByIdQuery, Customer>
    {
        private readonly ICustomerRepository _repository;

        public GetCustomerByIdQueryHandler(ICustomerRepository repository)
        {
            _repository = repository;
        }

        public async Task<Customer> Handle(GetCustomerByIdQuery request, CancellationToken cancellationToken)
        {
            var customer = await _repository.GetByIdAsync(request.Id);
            return customer;// ?? throw new Exception("Customer not found");
        }
    }
}
